TSK-1324: Moved tree from shared to administration component
