+++ web/src/app/administration/components/classification-details/classification-details.component.spec.ts
@@ -11,7 +11,7 @@ import { Location } from '@angular/common';
 import { MasterAndDetailService } from 'app/shared/services/master-and-detail/master-and-detail.service';
 import { RequestInProgressService } from 'app/shared/services/request-in-progress/request-in-progress.service';
 import { ClassificationsService } from 'app/shared/services/classifications/classifications.service';
-import { TreeNodeModel } from 'app/shared/models/tree-node';
+import { TreeNodeModel } from 'app/administration/models/tree-node';
 import { ImportExportService } from 'app/administration/services/import-export.service';
 import { EngineConfigurationSelectors } from 'app/shared/store/engine-configuration-store/engine-configuration.selectors';
 import { ClassificationSelectors } from 'app/shared/store/classification-store/classification.selectors';
